Relocated tests
Moved our tests to a new testclass: EqualsTest
